+++ src/err.jl
@@ -68,9 +68,9 @@ end
 export PyException
 
 ispy(x::PyException) = true
-getpy(x::PyException) = x.v
+Py(x::PyException) = x.v
 
-pyconvert_rule_exception(::Type{R}, x::Py) where {R<:PyException} = PyException(Py(x))
+pyconvert_rule_exception(::Type{R}, x::Py) where {R<:PyException} = pyconvert_return(PyException(x))
 
 function Base.show(io::IO, x::PyException)
     show(io, typeof(x))