+++ src/py_macro.jl
@@ -568,8 +568,8 @@ function py_macro_lower(st, body, ans, ex; flavour=:expr)
             tz = py_macro_lower(st, body3, ans, az)
             t = ty || tz
             if t
-                ty || push!(body2, :($ans = $Py($ans)))
-                tz || push!(body3, :($ans = $Py($ans)))
+                ty || push!(body2, :($ans = $pynew($Py($ans))))
+                tz || push!(body3, :($ans = $pynew($Py($ans))))
             end
             push!(body, Expr(:if, x, Expr(:block, body2...), Expr(:block, body3...)))
             return t
@@ -585,8 +585,8 @@ function py_macro_lower(st, body, ans, ex; flavour=:expr)
         ty = py_macro_lower(st, body2, ans, ay)
         t = tx || ty
         if t
-            tx || push!(body3, :($ans = $Py($ans)))
-            ty || push!(body2, :($ans = $Py($ans)))
+            tx || push!(body3, :($ans = $pynew($Py($ans))))
+            ty || push!(body2, :($ans = $pynew($Py($ans))))
         end
         push!(body, Expr(:if, :($pytruth($ans)), Expr(:block, body2...), Expr(:block, body3...)))
         return t
@@ -601,8 +601,8 @@ function py_macro_lower(st, body, ans, ex; flavour=:expr)
         ty = py_macro_lower(st, body3, ans, ay)
         t = tx || ty
         if t
-            tx || push!(body2, :($ans = $Py($ans)))
-            ty || push!(body3, :($ans = $Py($ans)))
+            tx || push!(body2, :($ans = $pynew($Py($ans))))
+            ty || push!(body3, :($ans = $pynew($Py($ans))))
         end
         push!(body, Expr(:if, :($pytruth($ans)), Expr(:block, body2...), Expr(:block, body3...)))
         return t
@@ -715,7 +715,7 @@ end
 
 function py_macro_lower_assign(st, body, lhs, rhs::Symbol)
     if lhs isa Symbol
-        push!(body, :($lhs = $Py($rhs)))
+        push!(body, :($lhs = $pynew($Py($rhs))))
     elseif @capture(lhs, ax_[ak__])
         @gensym x k
         tx = py_macro_lower(st, body, x, ax)
@@ -771,7 +771,7 @@ function py_macro(ex, mod, src)
             if v isa String
                 push!(inits, :($pycopy!($x, $pystr_intern!($pystr($v)))))
             else
-                push!(inits, :($pycopy!($x, $Py($v))))
+                push!(inits, :($pycopy!($x, $pynew($Py($v)))))
             end
         end
         push!(inits, :($doinit[] = false))