+++ src/pywrap/PyDict.jl
@@ -7,7 +7,7 @@ If `x` is not a Python object, it is converted to one using `pydict`.
 """
 struct PyDict{K,V} <: AbstractDict{K,V}
     py :: Py
-    PyDict{K,V}(x=pydict()) where {K,V} = new{K,V}(ispy(x) ? Py(x) : pydict(x))
+    PyDict{K,V}(x=pydict()) where {K,V} = new{K,V}(ispy(x) ? pynew(Py(x)) : pydict(x))
 end
 export PyDict
 
@@ -15,7 +15,7 @@ PyDict{K}(x=pydict()) where {K} = PyDict{K,Py}(x)
 PyDict(x=pydict()) = PyDict{Py,Py}(x)
 
 ispy(::PyDict) = true
-getpy(x::PyDict) = x.py
+Py(x::PyDict) = x.py
 pydel!(x::PyDict) = pydel!(x.py)
 
 pyconvert_rule_mapping(::Type{T}, x::Py, ::Type{PyDict{K,V}}=Utils._type_ub(T)) where {T<:PyDict,K,V} =