+++ src/pywrap/PyIterable.jl
@@ -5,14 +5,14 @@ This object iterates over iterable Python object `x`, yielding values of type `T
 """
 struct PyIterable{T}
     py :: Py
-    PyIterable{T}(x) where {T} = new{T}(Py(x))
+    PyIterable{T}(x) where {T} = new{T}(pynew(Py(x)))
 end
 export PyIterable
 
 PyIterable(x) = PyIterable{Py}(x)
 
 ispy(x::PyIterable) = true
-getpy(x::PyIterable) = x.py
+Py(x::PyIterable) = x.py
 pydel!(x::PyIterable) = pydel!(x.py)
 
 Base.IteratorSize(::Type{PyIterable{T}}) where {T} = Base.SizeUnknown()