+++ src/pywrap/PyList.jl
@@ -7,14 +7,14 @@ If `x` is not a Python object, it is converted to one using `pylist`.
 """
 struct PyList{T} <: AbstractVector{T}
     py :: Py
-    PyList{T}(x=pylist()) where {T} = new{T}(ispy(x) ? Py(x) : pylist(x))
+    PyList{T}(x=pylist()) where {T} = new{T}(ispy(x) ? pynew(Py(x)) : pylist(x))
 end
 export PyList
 
 PyList(x=pylist()) = PyList{Py}(x)
 
 ispy(::PyList) = true
-getpy(x::PyList) = x.py
+Py(x::PyList) = x.py
 pydel!(x::PyList) = pydel!(x.py)
 
 pyconvert_rule_sequence(::Type{T}, x::Py, ::Type{PyList{V}}=Utils._type_ub(T)) where {T<:PyList,V} =